feat: useSearchFishingSpot 훅 작성
